+++ src/main/cpp/auton/drivePrimitives/DriveStop.cpp
@@ -26,8 +26,8 @@
 #include "auton/drivePrimitives/DriveStop.h"
 #include "auton/drivePrimitives/IPrimitive.h"
 #include "auton/PrimitiveParams.h"
-#include "chassis/ChassisConfigMgr.h"
-#include "chassis/ChassisConfig.h"
+#include "chassis/configs/ChassisConfigMgr.h"
+#include "chassis/configs/ChassisConfig.h"
 #include "chassis/ChassisMovement.h"
 #include "configs/RobotConfig.h"
 #include "configs/RobotConfigMgr.h"
