+++ src/main/cpp/auton/drivePrimitives/ResetPositionPathPlanner.cpp
@@ -23,8 +23,8 @@
 #include "auton/drivePrimitives/IPrimitive.h"
 #include "auton/drivePrimitives/ResetPositionPathPlanner.h"
 #include "auton/PrimitiveParams.h"
-#include "chassis/ChassisConfig.h"
-#include "chassis/ChassisConfigMgr.h"
+#include "chassis/configs/ChassisConfig.h"
+#include "chassis/configs/ChassisConfigMgr.h"
 #include "chassis/SwerveChassis.h"
 #include "DragonVision/DragonVision.h"
 #include "utils/logging/Logger.h"
