+++ src/main/cpp/auton/drivePrimitives/ResetPositionPathPlannerNoVision.cpp
@@ -22,8 +22,8 @@
 #include "auton/drivePrimitives/IPrimitive.h"
 #include "auton/drivePrimitives/ResetPositionPathPlannerNoVision.h"
 #include "auton/PrimitiveParams.h"
-#include "chassis/ChassisConfig.h"
-#include "chassis/ChassisConfigMgr.h"
+#include "chassis/configs/ChassisConfig.h"
+#include "chassis/configs/ChassisConfigMgr.h"
 #include "chassis/SwerveChassis.h"
 #include "utils/logging/Logger.h"
 #include "utils/FMSData.h"
