+++ src/main/cpp/chassis/HolonomicDrive.cpp
@@ -26,8 +26,8 @@
 #include "chassis/ChassisMovement.h"
 #include "chassis/ChassisOptionEnums.h"
 #include "chassis/HolonomicDrive.h"
-#include "chassis/ChassisConfig.h"
-#include "chassis/ChassisConfigMgr.h"
+#include "chassis/configs/ChassisConfig.h"
+#include "chassis/configs/ChassisConfigMgr.h"
 #include "chassis/DragonDriveTargetFinder.h"
 #include "State.h"
 #include "teleopcontrol/TeleopControl.h"
