+++ src/main/cpp/chassis/configs/ChassisConfigChassis_9997.h
@@ -14,21 +14,48 @@
 //====================================================================================================================================================
 
 #pragma once
+#include "chassis/configs/ChassisConfig.h"
 
-#include "chassis/ChassisConfig.h"
-#include "configs/RobotConfigMgr.h"
+#include "units/length.h"
+#include "ctre/phoenix6/Pigeon2.hpp"
 
-class ChassisConfigMgr
+class ChassisConfigChassis_9997 : public ChassisConfig
 {
 public:
-	static ChassisConfigMgr *GetInstance();
-	ChassisConfig *GetCurrentConfig() const { return m_config; }
-	void InitChassis(RobotConfigMgr::RobotIdentifier id);
+	ChassisConfigChassis_9997() = default;
+	~ChassisConfigChassis_9997() = default;
+
+protected:
+	void DefinePigeon() override;
+	void DefineChassis() override;
 
 private:
-	ChassisConfigMgr();
-	~ChassisConfigMgr() = default;
+	std::string m_canbusName = std::string("rio");
+	const int m_leftfrontdriveID = 15;
+	const int m_leftfrontturnID = 14;
+	const double m_leftfrontOffset = 0.303466796875;
+	const bool m_leftfrontdriveInvert = true;
+	const bool m_leftfrontturnInvert = false;
+	const bool m_leftfrontcancoderInvert = false;
 
-	static ChassisConfigMgr *m_instance;
-	ChassisConfig *m_config;
+	const int m_leftbackdriveID = 1;
+	const int m_leftbackturnID = 0;
+	const double m_leftbackOffset = -0.32177734375;
+	const bool m_leftbackdriveInvert = true;
+	const bool m_leftbackturnInvert = false;
+	const bool m_leftbackcancoderInvert = false;
+
+	const int m_rightfrontdriveID = 13;
+	const int m_rightfrontturnID = 12;
+	const double m_rightfrontOffset = 0.30908203125;
+	const bool m_rightfrontdriveInvert = true;
+	const bool m_rightfrontturnInvert = false;
+	const bool m_rightfrontcancoderInvert = false;
+
+	const int m_rightbackdriveID = 3;
+	const int m_rightbackturnID = 2;
+	const double m_rightbackOffset = -0.364990234375;
+	const bool m_rightbackdriveInvert = true;
+	const bool m_rightbackturnInvert = false;
+	const bool m_rightbackcancoderInvert = false;
 };
