+++ src/main/cpp/chassis/driveStates/DriveToAmp.cpp
@@ -15,7 +15,7 @@
 
 // Team302 includes
 #include "chassis/driveStates/DriveToAmp.h"
-#include "chassis/ChassisConfigMgr.h"
+#include "chassis/configs/ChassisConfigMgr.h"
 #include "chassis/DragonDriveTargetFinder.h"
 #include "pathplanner/lib/path/PathPlannerTrajectory.h"
 #include "chassis/driveStates/RobotDrive.h"
