+++ src/main/cpp/chassis/driveStates/DriveToNote.cpp
@@ -24,7 +24,7 @@
 
 // Team302 Includes
 #include "DragonVision/DragonVision.h"
-#include "chassis/ChassisConfigMgr.h"
+#include "chassis/configs/ChassisConfigMgr.h"
 #include "chassis/driveStates/DriveToNote.h"
 #include "utils/FMSData.h"
 #include "DragonVision/DragonVisionStructs.h"
