+++ src/main/cpp/chassis/driveStates/RobotDrive.cpp
@@ -21,8 +21,8 @@
 #include "units/angle.h"
 
 // Team302 Includes
-#include "chassis/ChassisConfig.h"
-#include "chassis/ChassisConfigMgr.h"
+#include "chassis/configs/ChassisConfig.h"
+#include "chassis/configs/ChassisConfigMgr.h"
 #include "chassis/ChassisMovement.h"
 #include "chassis/driveStates/AntiTip.h"
 #include "chassis/driveStates/RobotDrive.h"
