+++ src/main/cpp/chassis/headingStates/SpecifiedHeading.cpp
@@ -16,8 +16,8 @@
 // Team302 Includes
 #include "chassis/ChassisOptionEnums.h"
 #include "chassis/headingStates/SpecifiedHeading.h"
-#include "chassis/ChassisConfig.h"
-#include "chassis/ChassisConfigMgr.h"
+#include "chassis/configs/ChassisConfig.h"
+#include "chassis/configs/ChassisConfigMgr.h"
 
 #include "utils/logging/Logger.h"
